Add files via upload
Its not letting me add the file bc its so big? I think Ill be able to add it to the onedrive. Heres the link to the datasets. I am taking all of the words between "Patient" and "Doctor" as the patients question. I believe for every patient question there is a doctor answer. This could be a really easy way to do it. I am struggling to get the doctor response right now but i think i get it.